// frontend/src/lib/api-contract.generated.ts
export interface components {
    schemas: {
// ...
        CreateBranchDto: {
            /** @description The name of the branch */
            name: string;
            /** @description The address of the branch */
            address: string;
            /** @description The phone number of the branch */
            phone?: string;
        };
        BranchCountResponseDto: {
            users?: number;
            orders?: number;
            inventories?: number;
        };
        BranchUserReferenceDto: {
            id: string;
            fullName: string;
            email: string;
            role: string;
        };
        BranchResponseDto: {
            id: string;
            name: string;
            address: string;
            phone?: string | null;
            isActive: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
            _count?: components["schemas"]["BranchCountResponseDto"];
            users?: components["schemas"]["BranchUserReferenceDto"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}